requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code formatting
black>=23.12.0
isort>=5.13.0

# Linting
pylint>=3.0.0
flake8>=7.0.0
mypy>=1.8.0

# Testing tools
pytest-mock>=3.12.0
pytest-xdist>=3.5.0

# Documentation
sphinx>=7.2.0
